Do not overwrite cridentials

+++ cridentials/install.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 cred_file=$HOME/.bash_cridentials
 
+if [ -f $cred_file ]; then
+    return
+fi
+
 cp ~/.gurubash/cridentials/bash_cridentials $cred_file
 
 # Configure the email and name (only once)
